Install router dom and change app.js
