+++ src/components/App.module.css
@@ -1,17 +1,5 @@
 .App {
   text-align: center;
+  display: grid;
 }
-
-.spin {
-  animation: App-logo-spin
-  infinite .2s linear;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
+
