+++ src/components/DetailPage/DetailPage.jsx
@@ -1,0 +1,11 @@
+import React, { Component } from 'react'
+
+export default class DetailPage extends Component {
+  render() {
+    return (
+      <div>
+        Hello World
+      </div>
+    )
+  }
+}
