+++ src/components/PokeList/PokeItem/PokeItem.module.css
@@ -9,7 +9,7 @@
   flex-direction: column;
   border-radius: 10px;
   outline: 2px;
-  column-gap: 5px;
+  line-height: 1pt;
   
 }
